requirements.txt
django-heroku
gunicorn
requests_oauthlib